+++ src/SWP/Bundle/ContentBundle/Doctrine/ORM/ImageRepository.php
@@ -16,15 +16,20 @@ declare(strict_types=1);
 
 namespace SWP\Bundle\ContentBundle\Doctrine\ORM;
 
-use Doctrine\ORM\QueryBuilder;
 use SWP\Bundle\ContentBundle\Doctrine\ImageRepositoryInterface;
-use SWP\Component\Common\Criteria\Criteria;
 use SWP\Bundle\StorageBundle\Doctrine\ORM\EntityRepository;
 
 class ImageRepository extends EntityRepository implements ImageRepositoryInterface
 {
-    public function getByCriteria(Criteria $criteria, array $sorting): QueryBuilder
+    /**
+     * {@inheritdoc}
+     */
+    public function findImageByAssetId(string $assetId)
     {
-        return $this->getQueryByCriteria($criteria, $sorting, 'i');
+        return $this->createQueryBuilder('i')
+            ->where('i.assetId = :assetId')
+            ->setParameter('assetId', $assetId)
+            ->getQuery()
+            ->getOneOrNullResult();
     }
 }